Moved types into folders
