+++ src/types/component-helpers/image.ts
